Kruskall - Wallis + Wilcoxon (Mann-Whitney U) and aov + Tukey-HSD tests for an ecogen object

Kruskall - Wallis + Wilcoxon (Mann-Whitney U) and aov + Tukey-HSD tests for an ecogen object

eco.pairtest(eco, df = c("P", "E", "A", "C"), x, test = c("wilcoxon",
  "tukey"), adjust = "fdr", only.p = TRUE, ...)

Arguments

eco
Object of class "ecogen".

df
The data frame for the analysis. Could be "P", "E" "A" or "C". For dominant data, "A" is here considered here a synonym of the G data frame.

x
The name of the S slot column with the groups for the analysis.

test
Test to perform ("wilcoxon", "tukey").

adjust
P-values correction method for multiple tests passed to p.adjust. Defalut is "fdr".

only.p
Should it be just a matrix with P-values returned? Default TRUE.

...
Additional arguments passed to wilcox.test or TukeyHSD.

Details

This program returns the Wilcoxon (Mann-Whitney U) or Tukey-HSD statistics and P-values for the multiple comparisons of the variables contained in the selected data frame, among the levels of a factor of the slot "S".

See also

wilcox.test TukeyHSD

Examples

# NOT RUN {
data(eco3)
wil <- eco.pairtest(eco = eco3, df = "P", x = "structure")
wil
wil <- eco.pairtest(eco = eco3,df = "E", x = "structure")
wil
wil <- eco.pairtest(eco = eco3, df = "P", x = "structure", only.p = FALSE)
wil
wil <- eco.pairtest(eco = eco3,df = "P", x = "structure", test = "tukey")
wil
# }