Plot for a connection network

Plot method for an eco.weight object. For examples, see eco.weight This function can make a static plot with the original coordinates and an additional graph with the coordinates transformed as ranks. It can also construct dynamic plots (force networks and circle networks).

eco.plotWeight(x, type = c("simple", "igraph", "edgebundle", "network"),
  group = NULL, fontSize = 10, ebColor = NULL, vertex.size = 10,
  vertex.label = NA, bounded = FALSE, ...)

Arguments

x
Connection network

type
Plot type: "edgebundle", for a circular network, "network" for a tension network

group
Vector with classes assigned to the individuals, in the same original order

fontSize
Argument passed to forceNetwork contained in the weight object (which is the order of the table used to construct the weights)

ebColor
Vector with edge colors for the groups of the edgebundler plot (Experimental feature)

vertex.size
Parameter to plot.igraph

vertex.label
Parameter passed to plot.igraph

bounded
Logical. Value to enable (TRUE) or disable (FALSE) the bounding box limiting the force network graph extent see forceNetwork.

...
Additional arguments passed to plot.igraph

Examples

# see the examples in the function eco.weight:
# ?eco.weight