rasterplot graph for eco.lsa results

Plot method for local spatial analysis

# S4 method for eco.multilsa
eco.rasterplot(x, grp = NULL, limits = NULL,
  title = NULL, z.name = NULL, vertical = TRUE, significant = TRUE,
  rescaled = FALSE, alpha = 0.05, interactivePlot = TRUE, ...)

Arguments

x
eco.multilsa object returned by eco.lsa or

grp
factor with groups to use in the plot. Default NULL

limits
values limits used for computing the data gradient for the plot

title
plot title

z.name
name for the legend

vertical
should be partitioned the populations on the x axis? Default TRUE.

significant
plot only significant results? Default TRUE

rescaled
plot the rescaled observed values ([-1,1] range)?

alpha
threshold P value for results with permutation tests. default = 0.05.

interactivePlot
Show an interactive plot via plotly? (default: TRUE)

...
additional arguments

See also

eco.lsa